Fix imports and some function locations
