+++ src/modules/chain-data/onchain.spec.ts
@@ -1,8 +1,8 @@
 import { PoolType, SubgraphPoolBase } from '@balancer-labs/sdk';
-import { Pool } from '../types';
+import { Pool } from '@/modules/pools';
 import { fetchTokens, fetchPoolsFromChain } from './onchain';
-import { subgraphPoolBase } from '../../test/factories/sor';
-import { poolBase } from '../../test/factories/pools';
+import { subgraphPoolBase } from '../../../tests/factories/sor';
+import { poolBase } from '../../../tests/factories/pools';
 
 jest.mock('@balancer-labs/sdk');
 jest.mock('@ethersproject/providers');
