+++ src/modules/pools/index.ts
@@ -1,3 +1,4 @@
 export * from './pool.service';
 export * from './pool.decorator';
 export * from './types';
+export * from './utils';
