+++ src/modules/pools/utils.ts
@@ -100,3 +100,14 @@ export function isSame(newPool: Pool, oldPool?: Pool): boolean {
   }
   return true;
 }
+
+
+export function getTokenAddressesFromPools(pools: Pool[]): string[] {
+  const tokenAddressMap = {};
+  pools.forEach(pool => {
+    pool.tokensList.forEach(address => {
+      tokenAddressMap[address] = true;
+    });
+  });
+  return Object.keys(tokenAddressMap);
+}
