+++ src/modules/prices/utils.ts
@@ -1,4 +1,5 @@
 import { Price } from '@balancer-labs/sdk';
+import { BigNumber } from 'bignumber.js';
 
 /** Formats a price correctly for storage. Does the following:
  *  - Converts prices in scientific notation to decimal (e.g. 1.63e-7 => 0.000000163)
@@ -7,7 +8,7 @@ import { Price } from '@balancer-labs/sdk';
 export function formatPrice(price: Price): Price {
   const formattedPrice: Price = {};
   Object.entries(price).forEach(([currency, value]) => {
-    formattedPrice[currency] = new OldBigNumber(value).toFixed();
+    formattedPrice[currency] = new BigNumber(value).toFixed();
   });
 
   return formattedPrice;
