+++ src/modules/server/server.spec.ts
@@ -1,7 +1,8 @@
 import { parseUnits } from '@ethersproject/units';
 import supertest from 'supertest';
 import { Network, TOKENS } from '../../constants';
-import { SorRequest, SerializedSwapInfo, Token } from '../../types';
+import { SorRequest, SerializedSwapInfo } from '@/modules/sor';
+import { Token } from '@/modules/tokens';
 import {
   createPoolsTable,
   createTokensTable,
@@ -12,11 +13,11 @@ import {
   updateToken,
   getPools,
   getTokens,
-} from './data-providers/dynamodb';
-import DBTOKENS from '../../../test/mocks/tokens.json';
-import DBPOOLS from '../../../test/mocks/pools';
-import server from '../../server';
-import { localAWSConfig } from '../../utils';
+} from '@/modules/dynamodb';
+import { localAWSConfig } from '@/modules/aws';
+import DBTOKENS from '../../../tests/mocks/tokens.json';
+import DBPOOLS from '../../../tests/mocks/pools';
+import server from './server';
 
 const AWS = require('aws-sdk');
 AWS.config.update(localAWSConfig);
