+++ src/modules/server/server.ts
@@ -1,12 +1,12 @@
 require('dotenv').config();
 import debug from 'debug';
 import express from 'express';
-import { getToken } from '../../modules/dynamodb/dynamodb';
-import { localAWSConfig } from '../../utils';
-import { handler as getPoolsHandler } from '../../lambdas/get-pools';
-import { handler as getPoolHandler } from '../../lambdas/get-pool';
-import { handler as sorHandler } from '../../lambdas/run-sor';
-import { handler as getTokensHandler } from '../../lambdas/get-tokens';
+import { getToken } from '@/modules/dynamodb';
+import { localAWSConfig } from '@/modules/aws';
+import { handler as getPoolsHandler } from '@/lambdas/get-pools';
+import { handler as getPoolHandler } from '@/lambdas/get-pool';
+import { handler as sorHandler } from '@/lambdas/run-sor';
+import { handler as getTokensHandler } from '@/lambdas/get-tokens';
 
 const log = debug('balancer');
