+++ src/modules/sor/types.ts
@@ -1,3 +1,4 @@
+import { SwapV2 } from "@balancer-labs/sdk";
 
 export interface SorRequest {
   sellToken: string;
