+++ tsconfig.json
@@ -14,6 +14,5 @@
       "@/*": ["./*"]
     }
   },
-  "include": ["src/**/*"],
-  "exclude": ["src/**/*.spec.ts"]
+  "include": ["src/**/*"]
 }
